+++ UniversityRegistrationSystem/src/model/person/Professor.java
@@ -8,7 +8,7 @@ public class Professor extends Person {
 	
 	public Professor(Address homeAddr, Address workAddr,
 			PhoneNumbers phoneNums, String fName, String lName,
-			int social, Date dob, String dept) {
+			long social, Date dob, String dept) {
 		super(homeAddr, workAddr, phoneNums, fName, lName,
 				social, dob);
 		department = dept;
